- merged successfully - excessive exceptional handling made for the validations on filds of the dto, request and response - front end pages like, home, courses, add course were made - split dev and prod mode was add .. now you can run the spring boot app alone and the other images alone to make modifications on the backend faster while the mysql and other services run on the background - check the pagitnations with the front ... first steps were made in the backend in ViewController.java - file storage system should be added to let all the user and courses have their images ... not hard coded images in the front end ... the link to the image should be saved in the back - todo the jwt expired exception should be handled properly eithr in the jwt filter or in the security exception handler (that doesn't even catch the thrown error)

+++ src/main/java/com/app/coursecenter/service/courseservice/CourseServiceImpl.java
@@ -1,16 +1,20 @@
 package com.app.coursecenter.service.courseservice;
 
 import com.app.coursecenter.entity.Course;
+import com.app.coursecenter.exception.DuplicateResourceException;
 import com.app.coursecenter.exception.ResourceNotFoundException;
 import com.app.coursecenter.mapper.CourseMapper;
 import com.app.coursecenter.repository.CourseRepository;
 import com.app.coursecenter.request.CreateCourseRequest;
 import com.app.coursecenter.request.UpdateCourseRequest;
+import com.app.coursecenter.response.CourseResponse;
+import com.app.coursecenter.service.FileStorageService.FileStorageService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -19,39 +23,52 @@ public class CourseServiceImpl implements CourseService {
     private final CourseRepository courseRepository;
     private final CourseMapper courseMapper;
 
+    // todo let a file storage system add the course image instead of using static resources
+    private final FileStorageService fileStorageService;
+
     @Override
-    public Page<CourseDto> getAllCourses(int page, int size) {
+    public Page<CourseResponse> getAllCourses(int page, int size) {
         Pageable pageable = PageRequest.of(page, size);
         Page<Course> coursePage = courseRepository.findAll(pageable);
-        return coursePage.map(courseMapper::courseToCourseDto);
+        return coursePage.map(courseMapper::courseToCourseResponse);
     }
 
     @Override
-    public Course getCourseById(Long id) {
-        return courseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Course Not found!"));
+    public CourseResponse getCourseById(Long id) {
+        return courseMapper.courseToCourseResponse(
+                courseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Course Not found!")));
     }
 
     @Override
-    public Course addCourse(CreateCourseRequest course) {
-        try {
-            return courseRepository.save(courseMapper.courseRequestToCourse(course));
-        } catch (Exception e) {
-            throw e;
+    public CourseResponse addCourse(CreateCourseRequest courseRequest) {
+
+        if (courseRepository.findCourseByName(courseRequest.getName()).isPresent()) {
+            throw new DuplicateResourceException("Course Already Exists by this name!");
         }
+
+
+        Course course = courseRepository.save(courseMapper.courseCreateRequestToCourse(courseRequest));
+        course.setRating(0); // the default of the Integer class is null
+        return courseMapper.courseToCourseResponse(course);
+
     }
 
     @Override
-    public Course updateCourse(Long id, CourseDto courseDto) {
-        return courseRepository.save(courseMapper.courseDtoToCourse(courseDto));
+    public CourseResponse updateCourse(Long id, UpdateCourseRequest courseUpdateRequest) {
+
+        // courseRequest maybe different from courseResponse in the future
+        Course course = courseRepository.save(courseMapper.courseUpdateRequestToCourse(courseUpdateRequest));
+        return courseMapper.courseToCourseResponse(course);
+
     }
 
     @Override
     public Boolean deleteCourse(Long id) {
-        courseRepository.deleteById(id);
+        if (getCourseById(id) != null) {
+            courseRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-//    @Override
-//    public Set<Course> getCoursesByStudentId(Long studentId) {
-//        return courseRepository.findCoursesByStudentId(studentId);
-//    }
 }
